broke out session identification into helper method
